use of 32-bit version for better splitting...?

+++ progs/cicada_extract_score_impl.hpp
@@ -850,7 +850,7 @@ struct PhrasePairSource
   typedef size_t    size_type;
   typedef ptrdiff_t difference_type;
 
-  typedef uint64_t                           hash_value_type;
+  typedef uint32_t                           hash_value_type;
   typedef utils::hashmurmur<hash_value_type> hasher_type;
 
   typedef boost::filesystem::path                            path_type;
